Continually refresh the property values.
Changed to QTreeView and custom PropertyModel and changed Property to record an id which we pass back to the TypeHandler to get the value. This way we can keep polling it. Also in future we should be able to write the values back.
In reality we only request the value when asked by the view, we simply tell the view every 100ms that the values have changed and it decides if it wants to re-read them.